pyproject.toml
[tool.poetry]
name = "watermark-app"
version = "0.1.0"
description = ""
authors = ["Andrew Best <andrew.best@tri.global>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
pillow = "^11.0.0"
pyqt5 = "^5.15.11"
# Poetry workaround to avoid failing deps
pyqt5-qt5 = "5.15.2"
pytest = "^8.3.4"
pytest-qt = "^4.4.0"
black = "^24.10.0"
pyinstaller = "^6.11.1"
pillow-heif = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
